// frontend/src/modules/common/types/footer.type.ts
export interface IColProps {
  width?: string;

  marginTop?: string;
}
export interface ITextProps {
  fontWeight: string;
  fontSize: string;
  marginTop?: string;
}
export interface LinkConteinerCol4Props {
  img: any;
  widthBef: string;
  heigthBef: string;
  marginLeft: string;
}
export interface LinkCompanyAndLinksProps {
  fontWeight: string;
}